reading in domainData
