+++ datadaddy_react/src/assets/data/domainMap.js
@@ -1,0 +1,44 @@
+export const domainData = [
+    {
+        "domain": "google.com",
+        "privacy_policy": "https://policies.google.com/privacy?hl=en",
+        "rtk_form": "https://takeout.google.com/settings/takeout",
+        "email": ""
+    },
+    {
+        "domain": "reddit.com",
+        "privacy_policy": "https://www.redditinc.com/policies/privacy-policy",
+        "rtk_form": "https://www.reddit.com/settings/data-request",
+        "email": "redditdatarequests@reddit.com"
+    },
+    {
+        "domain": "redditinc.com",
+        "privacy_policy": "https://www.redditinc.com/policies/privacy-policy",
+        "rtk_form": "https://www.reddit.com/settings/data-request",
+        "email": "redditdatarequests@reddit.com"
+    },
+    {
+        "domain": "facebook.com",
+        "privacy_policy": "https://www.facebook.com/legal/policy/ccpa",
+        "rtk_form": "https://www.facebook.com/help/contact/784491318687824",
+        "email": ""
+    },
+    {
+        "domain": "youtube.com",
+        "privacy_policy": "https://policies.google.com/privacy",
+        "rtk_form": "https://takeout.google.com/settings/takeout",
+        "email": ""
+    },
+    {
+        "domain": "twitter.com",
+        "privacy_policy": "https://twitter.com/en/privacy",
+        "rtk_form": "https://twitter.com/settings/download_your_data",
+        "email": ""
+    },
+    {
+        "domain": "",
+        "privacy_policy": "",
+        "rtk_form": "",
+        "email": ""
+    }
+]
