+++ datadaddy_react/src/assets/data/emailBoilerplate.txt
@@ -1,0 +1,18 @@
+To whom it may concern:
+
+I am writing to request access to personal information pursuant to Section 1798.110 of The California Consumer Privacy Act (CCPA). Please advise as to the following:
+1. Please confirm to me whether or not my personal information has been collected, sold or disclosed over the past 12 months. If so, please disclose:
+  1.1 What categories of personal information has been collected or disclosed for business purposes, and provide me with a copy of, or access to, my personal information that you have or are processing
+  1.2 Please identify the specific pieces of personal information that you have collected about me
+  1.3 Please advise what sources were used to obtain my personal information
+  1.4 Please advise what categories of my personal information that you have shared with or disclosed to third parties
+  1.5 Please advise in which countries my personal information is stored, or accessible from. In case you make use of cloud services to store or process my data, please include the countries in which the servers are located where my data are or were (in the past 12 months) stored
+2. Please provide me with a detailed accounting of the business or commercial purposes for which you are collecting or selling my personal information
+3. Please advise how long you store my personal information, and if retention is based upon the category of personal information, please identify how long each category is retained
+4. Please advise as to whether any categories of my personal information have been sold to a third party, and if so, what categories were included in such sale. If my personal information has been sold, please identify:
+  4.1 The categories of third parties to whom the information was sold
+  4.1 What specific personal information has been sold to such third party(ies)
+Please note that I do not consent to any personal information which is part of this request, including my email address and name, to be used for any purpose other than fulfilling this request.
+If you do not normally deal with these requests, please forward this email to the relevant person in your organization. Please note that you have 45 days to comply with this request as required under subsection 1798.130.
+
+Kind regards,
